objectified DeliveryData
