+++ packages/pymedphys_base/src/pymedphys_base/deliverydata/exportable.py
@@ -1,0 +1,202 @@
+# Copyright (C) 2019 Cancer Care Associates
+
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published
+# by the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version (the "AGPL-3.0+").
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License and the additional terms for more
+# details.
+
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+# ADDITIONAL TERMS are also included as allowed by Section 7 of the GNU
+# Affero General Public License. These additional terms are Sections 1, 5,
+# 6, 7, 8, and 9 from the Apache License, Version 2.0 (the "Apache-2.0")
+# where all references to the definition "License" are instead defined to
+# mean the AGPL-3.0+.
+
+# You should have received a copy of the Apache-2.0 along with this
+# program. If not, see <http://www.apache.org/licenses/LICENSE-2.0>.
+
+
+import functools
+from typing import Union, Tuple
+
+import numpy as np
+
+from .core import DeliveryDataBase
+from .utilities import find_relevant_control_points
+
+
+class DeliveryData(DeliveryDataBase):
+    def __new__(cls, *args, **kwargs):
+        return super().__new__(cls, *args, **kwargs)
+
+    @classmethod
+    def from_delivery_data_base(cls, delivery_data_base):
+        if type(delivery_data_base) is type(cls):
+            return delivery_data_base
+
+        return cls(*delivery_data_base)
+
+    @functools.lru_cache()
+    def filter_cps(self):
+        cls = type(self)
+
+        relevant_control_points = find_relevant_control_points(
+            self.monitor_units)
+
+        new_delivery_data = []
+        for item in self:
+            new_delivery_data.append(np.array(item)[relevant_control_points])
+
+        return cls(*new_delivery_data)
+
+    @functools.lru_cache()
+    def mask_by_gantry(self, angles: Union[Tuple, float, int], tolerance=3):
+        iterable_angles: tuple
+
+        try:
+            _ = iter(angles)  # type: ignore
+            iterable_angles = tuple(angles)  # type: ignore
+        except TypeError:
+            # Not iterable, assume just one angle provided
+            iterable_angles = (angles,)
+
+        return self.all_masked(self, iterable_angles, tolerance)
+
+    @functools.lru_cache()
+    def metersets(self, gantry_angles, gantry_tolerance):
+        all_masked_delivery_data = self.mask_by_gantry(
+            gantry_angles, gantry_tolerance)
+
+        metersets = []
+        for delivery_data in all_masked_delivery_data:
+            try:
+                metersets.append(delivery_data.monitor_units[-1])
+            except IndexError:
+                continue
+
+        return tuple(metersets)
+
+    def merge(self, *args):
+        separate = tuple(self,) + args
+        collection = {}  # type: ignore
+
+        for delivery_data in separate:
+            for field in delivery_data._fields:
+                try:
+                    collection[field] = np.concatenate(
+                        [collection[field], getattr(delivery_data, field)],
+                        axis=0
+                    )
+                except KeyError:
+                    collection[field] = getattr(delivery_data, field)
+
+        mu = np.concatenate([[0], np.diff(collection['monitor_units'])])
+        mu[mu < 0] = 0
+        collection['monitor_units'] = np.cumsum(mu)
+
+        for key, item in collection.items():
+            collection[key] = item.tolist()
+
+        merged = DeliveryDataBase(**collection)
+
+        return merged
+
+    def all_masked(self,
+                   template_gantry_angles, gantry_tol,
+                   quiet=False):
+        masks = self._gantry_angle_masks(
+            template_gantry_angles, gantry_tol, quiet=quiet)
+
+        all_masked_delivery_data = tuple(
+            self._apply_mask_to_delivery_data(mask)
+            for mask in masks
+        )
+
+        return all_masked_delivery_data
+
+    def extract_one_gantry_angle(self, gantry_angle, gantry_angle_tol=3):
+        near_angle = self._gantry_angle_mask(gantry_angle, gantry_angle_tol)
+
+        return self._apply_mask_to_delivery_data(near_angle)
+
+    def _gantry_angle_masks(self, gantry_angles,
+                            gantry_tol, quiet=False):
+        masks = [
+            self._gantry_angle_mask(gantry_angle, gantry_tol)
+            for gantry_angle in gantry_angles
+        ]
+
+        for mask in masks:
+            if np.all(mask == 0):
+                continue
+
+            # TODO: Apply mask by more than just gantry angle to appropriately
+            # extract beam index even when multiple beams have the same gantry
+            # angle
+            is_duplicate_gantry_angles = not np.sum(
+                np.abs(np.diff(np.concatenate([[0], mask, [0]])))) == 2
+
+            if is_duplicate_gantry_angles:
+                raise ValueError("Duplicate gantry angles not yet supported")
+
+        try:
+            assert np.all(np.sum(masks, axis=0) == 1), (
+                "Not all beams were captured by the gantry tolerance of "
+                " {}".format(gantry_tol)
+            )
+        except AssertionError:
+            if not quiet:
+                print("Allowable gantry angles = {}".format(gantry_angles))
+                gantry = np.array(self.gantry, copy=False)
+                out_of_tolerance = np.unique(
+                    gantry[np.sum(masks, axis=0) == 0]).tolist()
+                print("The gantry angles out of tolerance were {}".format(
+                    out_of_tolerance))
+
+            raise
+
+        return masks
+
+    def _gantry_angle_mask(self, gantry_angle, gantry_angle_tol):
+        near_angle = np.abs(
+            np.array(self.gantry) - gantry_angle) <= gantry_angle_tol
+        assert np.all(np.diff(np.where(near_angle)[0]) == 1)
+
+        return near_angle
+
+    def _apply_mask_to_delivery_data(self, mask):
+        cls = type(self)
+
+        new_delivery_data = []
+        for item in self:
+            new_delivery_data.append(np.array(item)[mask].tolist())
+
+        new_monitor_units = new_delivery_data[0]
+        try:
+            first_monitor_unit_item = new_monitor_units[0]
+        except IndexError:
+            return cls(*new_delivery_data)
+
+        new_delivery_data[0] = np.round(
+            np.array(new_delivery_data[0], copy=False)
+            - first_monitor_unit_item, decimals=7
+        ).tolist()
+
+        return cls(*new_delivery_data)
+
+    def _strip_delivery_data(self, skip_size):
+        cls = type(self)
+
+        new_delivery_data = []
+        for item in self:
+            new_delivery_data.append(np.array(item)[::skip_size].tolist())
+
+        return cls(*new_delivery_data)
