--- packages/pymedphys_deliverydata/src/pymedphys_deliverydata/__init__.py
@@ -1,3 +1,0 @@
-# Disable this for now
-
-# from .object import DeliveryData
