--- packages/pymedphys_deliverydata/src/pymedphys_deliverydata/_install_requires.py
@@ -1,8 +1,0 @@
-install_requires = [
-    "numpy >= 1.12, < 1.16",
-    "pymedphys_base >= 0.8.0dev17, < 0.9.0",
-    "pymedphys_dicom >= 0.8.0dev17, < 0.9.0",
-    "pymedphys_fileformats >= 0.8.0dev17, < 0.9.0",
-    "pymedphys_mudensity >= 0.8.0dev17, < 0.9.0",
-    "pymedphys_utilities >= 0.8.0dev17, < 0.9.0"
-]
