--- packages/pymedphys_deliverydata/src/pymedphys_deliverydata/_version.py
@@ -1,2 +1,0 @@
-version_info = [0, 8, 0, 'dev17']
-__version__ = "0.8.0dev17"
