--- packages/pymedphys_deliverydata/src/pymedphys_deliverydata/dicom/__init__.py
@@ -1,6 +1,0 @@
-from .from_dicom import (
-    delivery_data_from_dicom)
-
-from .to_dicom import (
-    delivery_data_to_dicom,
-    gantry_tol_from_gantry_angles)
