--- packages/pymedphys_deliverydata/src/pymedphys_deliverydata/logfile/__init__.py
@@ -1,1 +1,0 @@
-from .trf import delivery_data_from_logfile
