--- packages/pymedphys_deliverydata/src/pymedphys_deliverydata/object/__init__.py
@@ -1,1 +1,0 @@
-from .exportable import DeliveryData
