--- packages/pymedphys_utilities/src/pymedphys_utilities/rtplan/__init__.py
@@ -1,3 +1,0 @@
-from .controlpoints import (
-    find_relevant_control_points,
-    remove_irrelevant_control_points)
